+++ app/vmui/packages/vmui/src/state/customPanel/reducer.ts
@@ -1,24 +1,31 @@
 import { DisplayType, displayTypeTabs } from "../../pages/CustomPanel/DisplayTypeSwitch";
+import { getFromStorage, saveToStorage } from "../../utils/storage";
 import { getQueryStringValue } from "../../utils/query-string";
+import { SeriesLimits } from "../../types";
+import { DEFAULT_MAX_SERIES } from "../../constants/graph";
 
 export interface CustomPanelState {
   displayType: DisplayType;
   nocache: boolean;
   isTracingEnabled: boolean;
+  seriesLimits: SeriesLimits
 }
 
 export type CustomPanelAction =
   | { type: "SET_DISPLAY_TYPE", payload: DisplayType }
+  | { type: "SET_SERIES_LIMITS", payload: SeriesLimits }
   | { type: "TOGGLE_NO_CACHE"}
   | { type: "TOGGLE_QUERY_TRACING" }
 
 const queryTab = getQueryStringValue("g0.tab", 0) as string;
 const displayType = displayTypeTabs.find(t => t.prometheusCode === +queryTab || t.value === queryTab);
+const limitsStorage = getFromStorage("SERIES_LIMITS") as string;
 
 export const initialCustomPanelState: CustomPanelState = {
   displayType: (displayType?.value || "chart") as DisplayType,
   nocache: false,
   isTracingEnabled: false,
+  seriesLimits: limitsStorage ? JSON.parse(getFromStorage("SERIES_LIMITS") as string) : DEFAULT_MAX_SERIES
 };
 
 export function reducer(state: CustomPanelState, action: CustomPanelAction): CustomPanelState {
@@ -27,6 +34,12 @@ export function reducer(state: CustomPanelState, action: CustomPanelAction): Cus
       return {
         ...state,
         displayType: action.payload
+      };
+    case "SET_SERIES_LIMITS":
+      saveToStorage("SERIES_LIMITS", JSON.stringify(action.payload));
+      return {
+        ...state,
+        seriesLimits: action.payload
       };
     case "TOGGLE_QUERY_TRACING":
       return {